Add LogFileList component with story
